Moved controllers, middleware, validators to separate folders
